+++ accounts/admin.py
@@ -1,0 +1,3 @@
+from django.contrib import admin  # noqa: F401  # Ignore "imported but unused" for now
+
+# Register your models here.
