+++ accounts/models.py
@@ -1,0 +1,3 @@
+from django.db import models  # noqa: F401  # Ignore "imported but unused" for now
+
+# Create your models here.
