+++ parks/urls.py
@@ -1,7 +1,5 @@
 from django.urls import path
 from . import views
-from django.contrib.auth import views as auth_views
-from .views import register_view
 
 urlpatterns = [
     path("parks/", views.park_and_map, name="park_and_map"),
@@ -14,43 +12,6 @@ urlpatterns = [
     path("home/", views.home_view, name="home"),
     path("contact/", views.contact_view, name="contact"),
     path("", views.home_view, name="home"),
-    path("register/", register_view, name="register"),
-    path(
-        "login/",
-        auth_views.LoginView.as_view(template_name="parks/login.html"),
-        name="login",
-    ),
-    path("logout/", auth_views.LogoutView.as_view(), name="logout"),
-    path(
-        "password_reset/",
-        auth_views.PasswordResetView.as_view(
-            template_name="registration/password_reset_form.html",
-            email_template_name="registration/password_reset_email.html",
-            subject_template_name="registration/password_reset_subject.txt",
-        ),
-        name="password_reset",
-    ),
-    path(
-        "password_reset/done/",
-        auth_views.PasswordResetDoneView.as_view(
-            template_name="registration/password_reset_done.html"
-        ),
-        name="password_reset_done",
-    ),
-    path(
-        "reset/<uidb64>/<token>/",
-        auth_views.PasswordResetConfirmView.as_view(
-            template_name="registration/password_reset_confirm.html"
-        ),
-        name="password_reset_confirm",
-    ),
-    path(
-        "reset/done/",
-        auth_views.PasswordResetCompleteView.as_view(
-            template_name="registration/password_reset_complete.html"
-        ),
-        name="password_reset_complete",
-    ),
     path("api/checkin/", views.checkin_view, name="checkin"),
     path("api/bethere/", views.bethere_view, name="bethere"),
     path("messages/", views.all_messages_view, name="all_messages"),
