+++ pawpark/settings.py
@@ -25,7 +25,7 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 SECRET_KEY = "django-insecure-5eooghf%un4sb3)r4=d9%97$(wdb)4cdd75)1xzo3!5%4#@)@-"
 
 # SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = False
+DEBUG = True
 
 ALLOWED_HOSTS = [
     "pawpark-prod-env.eba-wdriaqjm.us-east-1.elasticbeanstalk.com",  # new
@@ -57,6 +57,7 @@ INSTALLED_APPS = [
     "moderation",
     "channels",
     "django.contrib.humanize",
+    "accounts",
 ]
 
 ASGI_APPLICATION = "pawpark.asgi.application"
@@ -75,8 +76,6 @@ if USE_REDIS:
     }
 else:
     CHANNEL_LAYERS = {"default": {"BACKEND": "channels.layers.InMemoryChannelLayer"}}
-
-
 
 
 MIDDLEWARE = [
@@ -183,7 +182,7 @@ STATICFILES_DIRS = [
 
 DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
 
-LOGIN_URL = "/login/"
+LOGIN_URL = "/accounts/login/"
 
 LOGIN_REDIRECT_URL = "home"
 LOGOUT_REDIRECT_URL = "home"
